pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "python"
version = "1.0.0"
description = "le_big_dump"
authors = [
	{ name = "Jonas Sundstrom", email = "jonas@kirilla.se" }
]
license = { file = "LICENSE" }
dependencies = [
	"argparse",
	"cryptography",
	"flask",
	"paramiko",
	"pyfiglet",
	"pymetasploit3",
	"python-docx",
	"python-nmap",
	"requests",
	"scapy",
]
classifiers = [
	# classifiers?
]

[project.entry-points.console_scripts]
arp_network_scanner = "arp_network_scanner.main:main"
curl2python = "weblogin_curl2python.main:main"
md5_hash_cracker = "md5_hash_cracker.main:main"
pickle_revshell_maker = "pickle_phone_home.pickler:main"
pickle_web_vuln = "pickle_phone_home.unpickler:main"
portscanner = "portscanner.main:main"
ssh_server_paramiko = "ssh_server_paramiko.main:main"
sshlogin = "sshlogin.main:main"
todo = "todo.todo:main"
web_directory_enumeration = "web_directory_enumeration.main:main"
web_downloader = "web_downloader.main:main"
web_subdomain_enumeration = "web_subdomain_enumeration.main:main"
weblogin = "weblogin.main:main"
weblogin_curl2python = "weblogin_curl2python.main:main"
weblogin_wordpress = "weblogin_worpress.main:main"
webmin_exploit = "webmin_exploit.main:main"
wordlist_maker = "wordlist_maker.wordlist_maker:main"
# port-scanning pipline
scan = "ports_dont_come_easy.scan:main"
probe = "ports_dont_come_easy.probe:main"
report = "ports_dont_come_easy.report:main"
